requirements.txt
fastapi
uvicorn
langchain
langgraph
pinecone-client
sentence-transformers
python-dotenv
pydantic